refactor : step2 리뷰 대응 - enum에 사칙연산 기능 추가
- https://github.com/next-step/java-racingcar/pull/560#discussion_r395710777
- 테스트 코드 재작성
